// src/Infrastructure/Enums.tsx
export enum EnumApiMethod {
    GET = 'GET',
    POST = 'POST',
    PUT = 'PUT',
    DELETE = 'DELETE',
}

export enum EnumAttributeType {
    MetaDataKeyPropHeading = 'propHeading',
    MetaDataKeyGridGroup = 'gridGroup',
    MetaDataKeyGridOrder = 'gridOrder'
}

export enum EnumDataDisplayGroup {
    Intro = 1,
    Temperature = 2,
    Wind = 3,
    Pressure = 4,
    Precipitation = 5,
    Visibility = 6,
    Condition = 7,
    Geography = 8,
    LocalTime = 9,
    Sun = 10,
    Moon = 11,
    Miscellaneous = 1000
}